Hacking the 5 round FS into the 3 round FS.
This is done in the dumbest possible fashion:
By computing the online transcript in the preprocessing phase,
then adding the transcripts to the Merkle tree described in KKW.
The root is added to the FS for the Cut-and-choose.

Membership is then shown in the MT for the online phase.

This means we are executing the online phase twice,
but avoid the loss of soundness.