Don't reserve 0 and 1
I'm on the fence about this. On the one hand, it's a nice way to bake constant functionality into the Bristol fashion format, which doesn't support it (but probably should). On the other, it makes Reverie less flexible. It's possible to easily handle on the circuit/witness generation side, so the loss of flexibility is a completely avoidable downside. But it still annoys me to have to add extra constant bits to every witness we generate.