Follow Bristol convention for implicit outputs
I don't expect we'll actually use this as it's rare that we get so lucky as to have all of our output wires in contiguous order without explicitly rewriting the wires or adding extra buffer gates for each output. Nonetheless, we can continue to use the explicit I/O convention so no harm done.