Provide an Implicit I/O Convention (#23)
* Update README with proposed changeset

* Automatically generate input gates from the header

This isn't how I wanted to implement this, but since our InstructionParser type doesn't currently support associated data (other than the instruction stream) this is the least-disruptive way of doing it.

When we have time to work on a larger refactor, we can update the FileStreamer and FileStream types to support metadata for the streamed data, and store information about the circuit interface there.

* Trim whitespace from parsed header wires

* Fix trim call order

* Don't reserve 0 and 1

I'm on the fence about this. On the one hand, it's a nice way to bake constant functionality into the Bristol fashion format, which doesn't support it (but probably should). On the other, it makes Reverie less flexible. It's possible to easily handle on the circuit/witness generation side, so the loss of flexibility is a completely avoidable downside. But it still annoys me to have to add extra constant bits to every witness we generate.

* Update circuit.txt to use new format

* Follow Bristol convention for implicit outputs

I don't expect we'll actually use this as it's rare that we get so lucky as to have all of our output wires in contiguous order without explicitly rewriting the wires or adding extra buffer gates for each output. Nonetheless, we can continue to use the explicit I/O convention so no harm done.

* Cargo clippy

* Add EQW and EQ

* Add Const to instruction type

* Implement const gates

* Cargo fmt, cargo clippy

* Remove constant share generators

* Fix tracing behavior

* Add const pattern to Reverie tests

Wasn't getting built from the companion

* Implement fromstr for the bristol header

* Extra clippy fix

* Eliminate some `unwrap`s

Annoying that this doesn't work for NoneError, but oh well

* Appropriately set masks to zero for constants

* Cargo fmt