Commit fieldswitched client to `master` (#35)
* Update README with proposed changeset

* Automatically generate input gates from the header

This isn't how I wanted to implement this, but since our InstructionParser type doesn't currently support associated data (other than the instruction stream) this is the least-disruptive way of doing it.

When we have time to work on a larger refactor, we can update the FileStreamer and FileStream types to support metadata for the streamed data, and store information about the circuit interface there.

* Trim whitespace from parsed header wires

* Fix trim call order

* Don't reserve 0 and 1

I'm on the fence about this. On the one hand, it's a nice way to bake constant functionality into the Bristol fashion format, which doesn't support it (but probably should). On the other, it makes Reverie less flexible. It's possible to easily handle on the circuit/witness generation side, so the loss of flexibility is a completely avoidable downside. But it still annoys me to have to add extra constant bits to every witness we generate.

* Update circuit.txt to use new format

* Follow Bristol convention for implicit outputs

I don't expect we'll actually use this as it's rare that we get so lucky as to have all of our output wires in contiguous order without explicitly rewriting the wires or adding extra buffer gates for each output. Nonetheless, we can continue to use the explicit I/O convention so no harm done.

* Cargo clippy

* Add EQW and EQ

* Add Const to instruction type

* Implement const gates

* Cargo fmt, cargo clippy

* Remove constant share generators

* Fix tracing behavior

* Add const pattern to Reverie tests

Wasn't getting built from the companion

* Implement fromstr for the bristol header

* Extra clippy fix

* Eliminate some `unwrap`s

Annoying that this doesn't work for NoneError, but oh well

* Appropriately set masks to zero for constants

* Cargo fmt

* Wait for at least 100MB of free RAM before scheduling async tasks

* Bump free memory requirement to 512MB

* Collect preprocessing commitments asychronously

* Wait for prover tasks to finish async-ly

* Wait for verifier tasks to finish asynchronously

* Wait for earlier preprocessor tasks to finish asynchronously

* Remove wait_for_mem

* Fix warnings in mod

* Fix overflow in chunk calculation

* Chunk preprocessing

* Switch option to usize in chunk calculation

* Chunk prover to fit in memory

* Chunk verifier to fit in memory

* Support chunks from companion

This introduces the unfortunate change that we now always need to load the entire circuit into memory. Yes, it scales slower than the in-memory netlists, but it could still cost us a few gigs of memory for large circuits.

* Fix chunk size calculation

* Consume inputs so that tasks release their memory on time

* Clippy fixes

* Lower usage per gate, increase base usage

Recent tests suggest this is a more accurate measure of our maximum memory usage

* Get field switched version compiling

* Remove unused enum

* Fix Clippy lints

Redundant clones and complex types

* Make field switching tests compile

* More clippy lints?

I updated rustc so maybe that's what did it

* Fix failing tests

* Add a fieldswitched client (#34)

* Make instructions serializable

* Use Bincode for the binary format

We can of course restore the old version if we really want to, but right now nothing targets it.

* Make regular scalars serializable too

* Add companion workspace

* Create field switched copy of companion

* Get fieldswitched version compiling

* Almost compiling with weird lifetimes

* Abandon the pretense of understanding generics

* Cargo fmt

* Get 64-bit arrays working

const generics wen

* left-in comment

* Bump version

* Remove workspaces so we actually run the tests

At some point we should fix this project's structure

* Add oneshot compilation target

* Fix imports

* needless returns

* Add a subtraction gate

* Add a test for a random program

* Rename oneshot items to match tests

* Cargo fmt

* Fix debug statements

* Remove vestigial comment

* Fix extra imports